+++ wxc_sdk/converged_recordings/__init__.py
@@ -1,6 +1,6 @@
 from collections.abc import Generator
 from datetime import datetime
-from typing import Optional, Union
+from typing import Optional, Union, List
 
 from dateutil.parser import isoparse
 from pydantic import Field
@@ -147,6 +147,7 @@ class RecordingStorageRegion(str, Enum):
     de = 'DE'
     au = 'AU'
     in_ = 'IN'
+    ca = 'CA'
 
 
 class ConvergedRecordingMeta(ApiModel):
@@ -156,7 +157,7 @@ class ConvergedRecordingMeta(ApiModel):
     owner_type: Optional[str] = None
     owner_name: Optional[str] = None
     owner_email: Optional[str] = None
-    storage_region: Optional[str] = None
+    storage_region: Optional[RecordingStorageRegion] = None
     service_type: Optional[str] = None
     version: Optional[str] = None
     service_data: Optional[RecordingServiceData] = None
@@ -201,8 +202,7 @@ class ConvergedRecordingsApi(ApiChild, base=''):
 
         The list returned is sorted in descending order by the date and time that the recordings were created.
 
-        Long result sets are split into `pages
-        <https://developer.webex.com/docs/basics#pagination>`_.
+        List recordings requires the spark-compliance:recordings_read scope.
 
         :param from_: Starting date and time (inclusive) for recordings to return, in any `ISO 8601
             <https://en.wikipedia.org/wiki/ISO_8601>`_ compliant format.
@@ -271,6 +271,8 @@ class ConvergedRecordingsApi(ApiChild, base=''):
 
         Only recordings of owner with the authenticated user may be retrieved.
 
+        Get Recording Details requires the spark-compliance:recordings_read scope.
+
         :param recording_id: A unique identifier for the recording.
         :type recording_id: str
         :rtype: :class:`ConvergedRecordingWithDirectDownloadLinks`
@@ -288,6 +290,8 @@ class ConvergedRecordingsApi(ApiChild, base=''):
 
         If a Compliance Officer deletes another user's recording, the recording will be inaccessible to regular users
         (host, attendees and shared), and to Compliance officer also. This action purges the recordings from Webex.
+
+        Delete a Recording requires the spark-compliance:recordings_write scope.
 
         :param recording_id: A unique identifier for the recording.
         :type recording_id: str
@@ -315,7 +319,9 @@ class ConvergedRecordingsApi(ApiChild, base=''):
         authenticated user.
 
         For information on the metadata fields, refer to `Metadata Guide
-        <https://developer.webex.com/docs/webex-calling-overview>`_
+        <https://developer.webex.com/docs/api/guides/consolidated-metadata-documentation-and-samples-guide>`_
+
+        Get Recording metadata requires the spark-compliance:recordings_read scope.
 
         :param recording_id: A unique identifier for the recording.
         :type recording_id: str
@@ -335,3 +341,45 @@ class ConvergedRecordingsApi(ApiChild, base=''):
         url = self.ep(f'convergedRecordings/{recording_id}/metadata')
         data = super().get(url, params=params)
         return ConvergedRecordingMeta.model_validate(data)
+
+    def reassign(self, reassign_owner_email: str, owner_email: str = None, recording_ids: List[str] = None):
+        """
+        Reassign Recordings
+
+        Reassigns recordings to a new user. As an administrator, you can reassign a list of recordings or all
+        recordings of a particular user to a new user.
+
+        The recordings can belong to an org user, a virtual line, or a workspace, but the destination user should only
+        be a valid org user.
+
+        * "reassignOwnerEmail" is always required. It must be provided in every case.
+
+        * "ownerEmail" and "recordingIds" cannot both be empty. At least one of them must be provided.
+
+        * If "recordingIds" is empty but "ownerEmail" is provided, it means that all recordings owned by the
+          "ownerEmail" will be reassigned to "reassignOwnerEmail."
+
+        * If "recordingIds" is provided and "ownerEmail" is also provided, only the recordings specified by
+          "recordingIds" that are owned by "ownerEmail" will be reassigned to "reassignOwnerEmail."
+
+        * If "ownerEmail" is empty but "recordingIds" is provided, the recordings specified by "recordingIds" will be
+          reassigned to "reassignOwnerEmail," regardless of the current owner.
+
+        The `spark-admin:recordings_write` scope is required to reassign recordings.
+
+        :param reassign_owner_email: New owner of the recordings.
+        :type reassign_owner_email: str
+        :param owner_email: Recording owner email. Can be a user, a virtual line, or a workspace.
+        :type owner_email: str
+        :param recording_ids: List of recording identifiers to be reassigned.
+        :type recording_ids: list[str]
+        :rtype: None
+        """
+        body = dict()
+        if owner_email is not None:
+            body['ownerEmail'] = owner_email
+        if recording_ids is not None:
+            body['recordingIds'] = recording_ids
+        body['reassignOwnerEmail'] = reassign_owner_email
+        url = self.ep('convergedRecordings/reassign')
+        super().post(url, json=body)